move benches to benches/
